BAU: Probably fix Service Bus proxy configuration (#26)
The options.WebProxy was being set when creating the client, but the
client is retrieved by `provider.GetRequiredService<ServiceBusClient>()`
of which the options were not being set on.

This changes the logic so that the ServiceBusClient which is added to
the DI by `AddServiceBusClient` has the WebProxy set on it if required
which is then used by `azureBuilder.AddClient`.